+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Include JSON files with pip distributions (#244)
 - Added flush to `civis_to_file` when passed a user-created buffer,
   ensuring the buffer contains the entire file when subsequently read.
+- Fix several tests in the `test_io` module (#248)
 
 ### Added
 - Added a utility function which can robustly split a Redshift schema name